% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{add_figure_1}
\alias{add_figure_1}
\title{Figure 1}
\usage{
add_figure_1(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{ggmatrix} object. The figure
}
\description{
Number of areas affected by psychopathology in children with and without
different subtypes of SLD.
}
\seealso{
\code{\link[GGally:ggmatrix]{GGally::ggmatrix()}}
}
