% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{add_figure_3}
\alias{add_figure_3}
\title{Figure 3}
\usage{
add_figure_3(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{ggplot} object. The figure
}
\description{
Trend in the prevalence of psychopathologies over \code{dsm5_cutoff_35_012}
}
