% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_2}
\alias{add_table_2}
\title{Table 2}
\usage{
add_table_2(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{flextable} object. The table
}
\description{
Numbers and percentages of children with different types of SLD
and average test scores.
}
