% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_3}
\alias{add_table_3}
\title{Table 3}
\usage{
add_table_3(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{flextable} object. The table
}
\description{
Numbers and percentages of children with psychopathology for
different types of SLD.
}
