% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_4}
\alias{add_table_4}
\title{Table 4}
\usage{
add_table_4(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{flextable} object. The table
}
\description{
Fisher’s exact test results.
}
