% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_all}
\alias{add_table_all}
\title{Add tables}
\usage{
add_table_all(manuscript, tab1, tab2, tab3, tab4)
}
\arguments{
\item{manuscript}{\code{officer} rdocx object. The manuscript}

\item{tab1}{\code{flextable} object. Table 1}

\item{tab2}{\code{flextable} object. Table 2}

\item{tab3}{\code{flextable} object. Table 3}

\item{tab4}{\code{flextable} object. Table 4}
}
\value{
\code{officer} rdocx object. The manuscript with added content
}
\description{
Add all tables and table captions to the manuscript.
}
