% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_suppl}
\alias{add_table_suppl}
\title{Add table to supplemental}
\usage{
add_table_suppl(supplemental, tab_suppl, text)
}
\arguments{
\item{supplemental}{\code{officer} rdocx object. The manuscript}

\item{tab_suppl}{\code{flextable} object. Supplemental table}

\item{text}{list. List of descriptive statistics strings}
}
\value{
\code{officer} rdocx object. The supplemental with added content
}
\description{
Add the table and table caption to the supplemental.
}
