% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text}
\alias{add_text}
\title{Add text blocks to manuscript}
\usage{
add_text(manuscript, text)
}
\arguments{
\item{manuscript}{\code{officer} rdocx object. The manuscript template}

\item{text}{list. List of named character strings to be added to
the manuscript}
}
\value{
\code{officer} rdocx object. The manuscript with placeholders replaced
}
\description{
Iterates over list of named character strings and replaces placeholders
("<<...>>") with corresponding names in the Microsoft Word template of the
manuscript.
}
\seealso{
\code{\link[officer:body_replace_all_text]{officer::body_replace_all_text()}}
}
