% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_correlation}
\alias{add_text_correlation}
\title{Range of correlation coefficients for manuscript text}
\usage{
add_text_correlation(data)
}
\arguments{
\item{data}{tbl. Correlation table}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Computes ranges of correlation coefficients for the tests assessing academic
performance as well as for the tests assessing psychopathological symptoms
and stores them as list of character strings, to be reported in the
manuscript.
}
