% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_descriptives}
\alias{add_text_descriptives}
\title{Descriptive statistics for manuscript text}
\usage{
add_text_descriptives(data)
}
\arguments{
\item{data}{tbl. Filtered data frame}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Computes all descriptive statistics to be reported in the
manuscript and stores them as list of character strings.
}
