% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_exclusion}
\alias{add_text_exclusion}
\title{Numbers of excluded participants for manuscript text}
\usage{
add_text_exclusion(data)
}
\arguments{
\item{data}{tbl. Transformed data frame}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Computes number of excluded participants to be reported in the
manuscript and stores them as list of character strings.
}
