% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_fisher}
\alias{add_text_fisher}
\title{Results of Fisher's exact test for manuscript text}
\usage{
add_text_fisher(data)
}
\arguments{
\item{data}{tbl. Data frame with results of Fisher's exact
test}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Stores results of Fisher's exact test as list of
character strings.
}
