% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_poisson}
\alias{add_text_poisson}
\title{Results of the poisson model for manuscript text}
\usage{
add_text_poisson(data)
}
\arguments{
\item{data}{tbl. Data frame with results of the poisson model}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Stores results of the poisson model as list of
character strings.
}
