% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_posthoc}
\alias{add_text_posthoc}
\title{Results of the post hoc tests for manuscript text}
\usage{
add_text_posthoc(data)
}
\arguments{
\item{data}{tbl. Data frame with results of the posthoc test}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Stores results of the post hoc tests as list of
character strings.
}
