% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{add_text_trend}
\alias{add_text_trend}
\title{Results of trend test for manuscript text}
\usage{
add_text_trend(data)
}
\arguments{
\item{data}{tbl. Data frame with results of the trend test}
}
\value{
list. A list with named entries for the character
strings
}
\description{
Stores results of the trend test as list of
character strings.
}
