% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{adj_inv_sinh_ci_or}
\alias{adj_inv_sinh_ci_or}
\title{Confidence interval for odds ratios}
\usage{
adj_inv_sinh_ci_or(data, x, y, psi1 = 0.6, psi2 = 0.4, alpha = 0.05)
}
\arguments{
\item{data}{tbl. Input data}

\item{x, y}{character. Variable names}

\item{psi1, psi2}{double. Pseudo-frequencies (both should be > 0),
Default: 0.6, 0.4}

\item{alpha}{double. The nominal level, Default: 0.05}
}
\value{
tbl. Data frame with odds ratio and CIs
}
\description{
Computes the adjusted inverse hyperbolic sine confidence interval for the
odds ratio.
}
\details{
Procedure as described in Fagerland & Newcombe (2013) and Fagerland,
Lydersen, & Laake (2017).

The function is an adaptation of the \code{Matlab} script provided by Fagerland et
al. (2017) on the book's
\href{http://contingencytables.com/software-resources}{website}.
}
\references{
Fagerland, M., Lydersen, S., & Laake, P. (2017). \emph{Statistical analysis
of contingency tables.} London: Chapman and Hall.

Fagerland, M. W., & Newcombe, R. G. (2013). Confidence intervals for
odds ratio and relative risk based on the inverse hyperbolic sine
transformation. \emph{Statistics in Medicine, 32}(16), 2823–2836.
}
