% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{body_add_caption_figure}
\alias{body_add_caption_figure}
\title{Figure caption}
\usage{
body_add_caption_figure(manuscript, number, text)
}
\arguments{
\item{manuscript}{\code{officer} rdocx object. The manuscript}

\item{number}{character. Number of the figure}

\item{text}{character. Text of the caption}
}
\value{
\code{officer} rdocx object. The manuscript with added content
}
\description{
Adds a figure caption to the manuscript.
}
