% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{body_add_caption_table}
\alias{body_add_caption_table}
\title{Table caption}
\usage{
body_add_caption_table(manuscript, number, text)
}
\arguments{
\item{manuscript}{\code{officer} rdocx object. The manuscript}

\item{number}{character. Number of the table}

\item{text}{character. Text of the caption}
}
\value{
\code{officer} rdocx object. The manuscript with added content
}
\description{
Adds a table caption to the manuscript.
}
