% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{combine_wrt}
\alias{combine_wrt}
\title{Combine grade-specific WRT scores}
\usage{
combine_wrt(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Create variables \code{wrt_*} combinining the separate variables
for grade 3 and 4.
}
