% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{compare_dirs}
\alias{compare_dirs}
\title{Compare directories}
\usage{
compare_dirs(to = getwd())
}
\arguments{
\item{to}{character. File path, where to copy analysis structure,
Default: \code{getwd()}}
}
\description{
Compares analysis directories in source and target directory.
}
