% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{copy_analysis}
\alias{copy_analysis}
\title{Copy analysis directory structure}
\usage{
copy_analysis(to = getwd(), drake = TRUE, overwrite = FALSE)
}
\arguments{
\item{to}{character. File path, where to copy analysis structure,
Default: \code{getwd()}}

\item{drake}{logical. Whether to copy \code{.drake} directory, Default: \code{TRUE}}

\item{overwrite}{logical. Whether to overwrite the analysis directory
structure, Default: \code{FALSE}}
}
\description{
Copies the analysis directory structure of a packaged \code{drake}
workflow to the selected directory.
}
\details{
This function will copy the analysis directory structure supplied with
the package and the \code{.drake/} directory into the user's working directory
(or the directory supplied to the \code{to} parameter).
}
