% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{correct_mult_comp}
\alias{correct_mult_comp}
\title{Multiple comparison correction}
\usage{
correct_mult_comp(data, var_name, threshold = 0.05)
}
\arguments{
\item{data}{tbl. A tidy data frame with statistical test results}

\item{var_name}{character (vector). Column name(s) with p-values}

\item{threshold}{double. False discovery rate, Default: 0.05}
}
\value{
tbl. The input data frame with an additional column indicating
whether a p-value survived multiple comparison correction
}
\description{
Applies multiple comparison correction by controlling the false discovery
rate (FDR) using the modified FDR procedure by Benjamini and
Yekutieli (2001)
}
\references{
Benjamini, Y., & Yekutieli, D. (2001). The control of the false discovery
rate in multiple testing under dependency. \emph{The Annals of Statistics, 29}
(4), 1165–1188.
}
\seealso{
\code{\link[stats:p.adjust]{stats::p.adjust()}}
}
