% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{count_filter}
\alias{count_filter}
\title{Number of excluded participants}
\usage{
count_filter(data, criterion)
}
\arguments{
\item{data}{tbl. The data frame to be filtered}

\item{criterion}{character. The exclusion criterion}
}
\value{
double. The number of excluded participants
}
\description{
Computes the number of participants that are excluded when applying the
specified exclusion criterion.
}
