% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_nationality}
\alias{create_nationality}
\title{Create variable \code{nationality}}
\usage{
create_nationality(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Create variable indicating whether a child has German
or other nationality.
}
