% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_psy_vars}
\alias{create_psy_vars}
\title{Create variables indicating psychopath. problems}
\usage{
create_psy_vars(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Create variables indicating children's psychopathological problems.
A z-score >= 1 is used to classify children as having problems and a
z-score < 1 as not having problems.
}
\details{
\code{{psychopath}_cat}: group children as having / not having specific
psychopath. problems.

\code{psychopaths_n}: counts number of areas in which a child has
psychopath. problems.
}
