% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_sld_vars}
\alias{create_sld_vars}
\title{Create variables indicating SLD status}
\usage{
create_sld_vars(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Create variables \code{dsm5_cutoff_35{_*}} indicating SLD status according to
DSM-V. A cutoff of z-score <= -1.5 (T-score <= 35) is used to classify
children as having an SLD and a z-score > -1 to classify children as not
having an SLD.
}
\details{
\code{dsm5_cutoff_35}: groups children as having no SLD, an isolated SLD in
reading, spelling, or arithmetic or the different combinations of the
three SLDs (8 levels).

\code{dsm5_cutoff_35_{read/spell/math}}: group children as having or not
having a certain SLD.

\code{dsm5_cutoff_35_01}: groups children as having or not having any SLD.

\code{dsm5_cutoff_35_012}: groups children as having no, an isolated SLD, or
comorbid SLDs.

\code{dsm5_cutoff_35_n}: counts number of SLDs a child has.
}
