% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{delete_analysis}
\alias{delete_analysis}
\title{Delete analysis directory structure}
\usage{
delete_analysis(to = getwd())
}
\arguments{
\item{to}{character. File path, where to delete the directories from,
Default: \code{getwd()}.}
}
\description{
Deletes the analysis directory structure of a packaged \code{drake}
workflow from a selected directory.
}
\details{
This function will delete a previously copied analysis directory
structure from the user's working directory (or the directory supplied to
the \code{to} parameter).
}
