% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{export_pkgenv}
\alias{export_pkgenv}
\title{Export package environment}
\usage{
export_pkgenv()
}
\value{
The package environment
}
\description{
Exports the package environment, which holds several variables like the
name of the package, authors, title, etc. for use in other functions.
}
