% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{export_tiff_figure}
\alias{export_tiff_figure}
\title{Export `.tiff`` figure}
\usage{
export_tiff_figure(the_plot, name, width, height, dpi = 300)
}
\arguments{
\item{the_plot}{plot object. The plot to save}

\item{name}{character. The file name to save the plot under}

\item{width}{numeric. Width in cm}

\item{height}{numeric. Height in cm}

\item{dpi}{numeric. Resolution in DPI}
}
\description{
Export figure as \code{.tiff} file with appropriate size and resolution as
required by the journal
}
