% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(data, df_filter_cond)
}
\arguments{
\item{data}{tbl. The data frame to be filtered}

\item{df_filter_cond}{tbl. A data frame with exclusion criteria}
}
\value{
tbl. The filtered data frame
}
\description{
Filter out participants that fulfill one or several the exclusion criteria.
}
