% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fisher_test}
\alias{fisher_test}
\title{Fisher's exact test}
\usage{
fisher_test(data, x, y, alternative = "greater")
}
\arguments{
\item{data}{tbl. Input data}

\item{x, y}{character. Variable names}

\item{alternative}{character. Alternative hypothesis, Default: "greater"}
}
\value{
tbl. Data frame with p-value for the test, odds ratio, and adjusted
inverse hyperbolic sine CIs
}
\description{
Computes Fisher's exact test for 2 x 2 contingency tables.
}
\seealso{
\code{\link[=adj_inv_sinh_ci_or]{adj_inv_sinh_ci_or()}}
}
