% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{get_file_list}
\alias{get_file_list}
\title{Get file list}
\usage{
get_file_list(dir)
}
\arguments{
\item{dir}{character. File path, where the files are searched for}
}
\description{
Gets list of all files and directories in the selected directory
that are part of the analysis directory structure.
}
