% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{get_mean_ci}
\alias{get_mean_ci}
\title{Mean and CI of intelligence score}
\usage{
get_mean_ci(data, filter_cond)
}
\arguments{
\item{data}{tbl. Input data}

\item{filter_cond}{character. Filter condition(s)}
}
\value{
character. Mean (CI) of IQ value for selected subgroup
}
\description{
Computes the mean and confidence intervall of the IQ score for a
selected subgroup of participants.
}
