% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_perc_filter}
\alias{get_n_perc_filter}
\title{Compute number and percentage}
\usage{
get_n_perc_filter(data, filter, invert = FALSE)
}
\arguments{
\item{data}{tbl. A data frame to be filtered}

\item{filter}{character. Filter condition(s)}

\item{invert}{logical. Should the the filter condition(s) be inverted,
Default: FALSE}
}
\value{
numeric vector \code{c(number, percent)}. The number and percentage of
cases fulfilling the filter condition(s)
}
\description{
Computes the number and percentage of cases that fulfil
a certain filter condition
}
