% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_percent}
\alias{get_percent}
\title{Compute percentage}
\usage{
get_percent(count, total)
}
\arguments{
\item{count}{numeric. Number of selected cases}

\item{total}{numeric. Total number of cases}
}
\value{
double. Percentage rounded to one decimal
}
\description{
Computes the percentage of a selected number of cases
among the total number of cases.
}
