% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{get_plan}
\alias{get_plan}
\title{Get the analysis plan}
\usage{
get_plan()
}
\value{
A \code{drake} plan.
}
\description{
Using the \code{drake} pipeline toolkit, sets up all analysis steps to go from
the raw data to the finished manuscript.
}
\seealso{
\code{\link[drake:drake]{drake::drake()}}
}
