% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{glm_poisson}
\alias{glm_poisson}
\title{Poisson regression}
\usage{
glm_poisson(data)
}
\arguments{
\item{data}{Input data}
}
\value{
tbl. A tidy data frame with p-value, coefficients, and CIs for
the model
}
\description{
Computes a general linear model with logit link (poisson regression)
with number of SLDs as predictor and the number of psychopathological
areas as outcome variable.
}
\seealso{
\code{\link[stats:glm]{stats::glm()}} \code{\link[=tidy_glm]{tidy_glm()}}
}
