% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_figure_1}
\alias{plot_figure_1}
\title{Plot for Figure 1}
\usage{
plot_figure_1(data, labels)
}
\arguments{
\item{data}{tbl. Input data}

\item{labels}{list. A list containing, for each SLD to plot, the variable
name, the filter condition to filter the data frame and compute
percentages, and the name to use in the plot legend}
}
\value{
\code{ggplot} object. The plot
}
\description{
Plots line graphs representing percent of participants having a
certain SLD over the number of areas with psychopathological
problems
}
