% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_figure_2}
\alias{plot_figure_2}
\title{Plot for Figure 2}
\usage{
plot_figure_2(data, filter_var, invert_filter = FALSE, ymax = NULL)
}
\arguments{
\item{data}{tbl. Input data}

\item{filter_var}{character. The SLD variable name}

\item{invert_filter}{logical. Whether to select children \emph{without} the SLD,
Default: FALSE}

\item{ymax}{double. The maximum value of the y-axis of the intersection
size bar plot, Default: NULL}
}
\value{
\code{ggplot} object. The plot
}
\description{
Plots "Upset" graph visualizing the overlap between different
psychopathological problems within a selected subgroup of participants
having / not having a specific SLD
}
\seealso{
\code{\link[UpSetR:upset]{UpSetR::upset()}}
}
