% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_figure_3}
\alias{plot_figure_3}
\title{Plot for Figure 3}
\usage{
plot_figure_3(data, disability, psychopaths)
}
\arguments{
\item{data}{tbl. Input data}

\item{disability}{character. SLD variable name}

\item{psychopaths}{character vector. Vector of psychopathology variable names}
}
\value{
\code{ggplot} object. The plot
}
\description{
Plots line graphs representing percentages of selected psychopathologies
over the levels of an SLD variable.
}
