% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{posthoc_test}
\alias{posthoc_test}
\title{Post hoc tests for a 3 x 2 contingency table trend test}
\usage{
posthoc_test(data, x, y)
}
\arguments{
\item{data}{tbl. Input data}

\item{x}{character. Name of a categorical variable with r levels}

\item{y}{character. Name of a categorical variable with 2 levels}
}
\value{
tbl. Data frame with p-value for the test, odds ratio, and adjusted
inverse hyperbolic sine CIs for both comparisons
}
\description{
Computes one-sided Fisher’s exact tests comparing increasing levels of
the categorical variable with 3 levels (1 vs. 2; 2 vs. 3)
}
\seealso{
\code{\link[=fisher_test]{fisher_test()}} \code{\link[=adj_inv_sinh_ci_or]{adj_inv_sinh_ci_or()}}
}
