% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{recode_analyze}
\alias{recode_analyze}
\title{Recode variable \code{analyze}}
\usage{
recode_analyze(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Recode \code{analyze} to 1 for all children without siblings.
}
