% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{recode_education_mother}
\alias{recode_education_mother}
\title{Recode variable \code{education_mother}}
\usage{
recode_education_mother(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Recode \code{education_mother} to 3 levels instead of 5 by combining
the highest two school degrees as well as
no degree and the lowest school degree
}
