% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_fisher_or}
\alias{report_fisher_or}
\title{Odds ratio and CI for Fisher's exact test}
\usage{
report_fisher_or(data, sld, psy)
}
\arguments{
\item{data}{tbl. Data frame with results of Fisher's exact
test}

\item{sld}{character. SLD variable name}

\item{psy}{character. Psychopathology variable name}
}
\value{
character. String reporting odds ratio and CI,
e.g., "1.25 (95% CI = 1.15–1.35)"
}
\description{
Outputs a character string with the odds ratio and CI for
a specified combination of SLD and psychopathology
}
