% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_fisher_or_range}
\alias{report_fisher_or_range}
\title{Range of odds ratios from Fisher's exact test for specific psychopathology}
\usage{
report_fisher_or_range(data, psy)
}
\arguments{
\item{data}{tbl. Data frame with results of Fisher's exact
test}

\item{psy}{character. Psychopathology variable name}
}
\value{
character. String reporting odds ratio range,
e.g., "2.5–3.5"
}
\description{
Outputs a character string with the range of odds ratios for all
combinations of a specific psychopathology and the different SLDs
}
