% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_poisson}
\alias{report_poisson}
\title{Parameter estimate, CI, and p-value for the poisson model}
\usage{
report_poisson(data, the_term)
}
\arguments{
\item{data}{tbl. Data frame with results of the poisson model}

\item{the_term}{character. Model term}
}
\value{
character. Strings reporting parameter estimate, CI, and
p-value, e.g., "1.65 (95-CI = 1.55–1.75, p < .001)"
}
\description{
Outputs a character string with the estimate, CI, and p-value
for a specified term in the poisson model
}
