% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_posthoc}
\alias{report_posthoc}
\title{Odds ratios, CIs, and p-values for post hoc tests}
\usage{
report_posthoc(data, psy)
}
\arguments{
\item{data}{tbl. Data frame with results of the posthoc test}

\item{psy}{character. Psychopathology variable name}
}
\value{
character vector. Vector of strings reporting odds ratios,
p-value and CIs, e.g., "OR = 1.25, 95%-CI = 1.15–1.35, p < 0.001"
}
\description{
Outputs a character string with the odds ratios, CIs, and
p-values for a specified psychopathology
}
