% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_range_corr}
\alias{report_range_corr}
\title{Range of correlation coefficients}
\usage{
report_range_corr(data, vars)
}
\arguments{
\item{data}{tbl. Correlation table}

\item{vars}{character vector. Selection of variables to compute range for}
}
\value{
character. String reporting range of correlations, e.g., ".51–.8"
}
\description{
Outputs a character string with the range of correlation coefficients within
a specified selection of variables.
}
