% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{report_trend_sign}
\alias{report_trend_sign}
\title{z-score and p-value for trend test}
\usage{
report_trend_sign(data, psy)
}
\arguments{
\item{data}{tbl. Data frame with results of the trend test}

\item{psy}{character. Psychopathology variable name}
}
\value{
character. String reporting z-score and p-value,
e.g., "z= 4.45, p < 0.001"
}
\description{
Outputs a character string with z-score and p-value
for a specified psychopathology
}
