% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis}
\alias{reproduce_analysis}
\title{Reproduce the analysis}
\usage{
reproduce_analysis(to = getwd(), overwrite = FALSE, re_run = FALSE)
}
\arguments{
\item{to}{character. File path, where to copy analysis structure,
Default: \code{getwd()}}

\item{overwrite}{logical. Whether to overwrite the analysis directory
structure, Default: \code{FALSE}}

\item{re_run}{logical. Whether to re-run the analysis from scratch.}
}
\description{
Copies bundled analysis directory structure to selected directory
and reproduces analysis.
}
\details{
This function will copy the analysis directory structure supplied
with the package and the \code{.drake/} directory into the user's working
directory (or the directory supplied to the \code{to} parameter) and reproduce
the analysis by running \code{drake::make()}.
}
