% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tidy_glm}
\alias{tidy_glm}
\title{Tidy GLM results}
\usage{
tidy_glm(glm_model, exp = TRUE, abbr = FALSE)
}
\arguments{
\item{glm_model}{A GLM model object}

\item{exp}{Logical. Should coefficients and CIs be exponentiated for
binomial or poisson models, Default: TRUE}
}
\value{
tbl. A data frame with p-value, coefficients, and CIs
}
\description{
Return a tidy data frame of GLM results.
}
\details{
Wraps around \code{\link[broom:tidy]{broom::tidy()}} and \code{\link[broom:confint_tidy]{broom::confint_tidy()}} to provide a
concise overview of model parameters. Can exponentiate coefficients
and CIs for binomial or poisson models.
}
\seealso{
\code{\link[broom:tidy]{broom::tidy()}} \code{\link[broom:confint_tidy]{broom::confint_tidy()}}
}
