% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_data}
\alias{transform_data}
\title{Transform raw data}
\usage{
transform_data(data)
}
\arguments{
\item{data}{tbl. The data frame to be transformed}
}
\value{
tbl. The transformed data frame
}
\description{
Applies all transformations to the raw data.
}
