% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{trend_test}
\alias{trend_test}
\title{Trend test for ordered r x 2 contingency table}
\usage{
trend_test(data, x, y, alpha = 0.05)
}
\arguments{
\item{data}{tbl. Input data}

\item{x}{character. Name of a categorical variable with r levels}

\item{y}{character. Name of a categorical variable with 2 levels}

\item{alpha}{double. The nominal level, Default: 0.05}
}
\value{
tbl. Data frame with Wald test statistic and p-value as well as
estimate for the trend (odds ratio) and associated Wald confidence
intervals
}
\description{
Computes trend test based on the generalized linear model with logit link
function (logistic regression) and the Wald test statistic.
}
\details{
Procedure as described in Fagerland, Lydersen, & Laake (2017).

The function is an adaptation of the \code{Matlab} script provided by Fagerland et
al. (2017) on the book's
\href{http://contingencytables.com/software-resources}{website}.
}
\references{
Fagerland, M., Lydersen, S., & Laake, P. (2017). \emph{Statistical analysis
of contingency tables.} London: Chapman and Hall.
}
