% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{add_figure_all}
\alias{add_figure_all}
\title{Add figures}
\usage{
add_figure_all(manuscript)
}
\arguments{
\item{manuscript}{\code{officer} rdocx object. The manuscript}
}
\value{
\code{officer} rdocx object. The manuscript with added content
}
\description{
Add all figure captions to the manuscript.
}
