% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_1}
\alias{add_table_1}
\title{Table 1}
\usage{
add_table_1(data)
}
\arguments{
\item{data}{tbl. The filtered data frame}
}
\value{
\code{flextable} object. The table
}
\description{
Numbers and Percentages of Children with Anxiety, Depression, Conduct
Disorder, and ADHD in Children with Different Types of SLD
}
