% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_1}
\alias{add_table_1}
\title{Table 1}
\usage{
add_table_1(data)
}
\arguments{
\item{data}{tbl. The correlation table}
}
\value{
\code{flextable} object. The table
}
\description{
Correlation Coefficients between the Learning Ability Scores in Reading,
Spelling, and Arithmetic, and the Scores for Psychopathological Symptoms of
ADHD, Anxiety, Conduct Disorder, and Depression
}
