% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_table_sup}
\alias{add_table_sup}
\title{Supplemental Table}
\usage{
add_table_sup(data)
}
\arguments{
\item{data}{tbl. The correlation table}
}
\value{
\code{flextable} object. The table
}
\description{
Average Test Scores for the Different SLD Groups
}
